backend/requirements.txt
Flask==3.0.3
PyPDF2==3.0.1
python-docx==1.1.2
ollama==0.1.8